# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shirt Store API",
        "description": "No-login storefront API for a retro football shirt shop. Browse the catalog and shipping wilayas, then place an order without an account. All amounts are in DZD. Admin endpoints require the X-Admin-Token header.",
        "contact": {
            "email": "hello@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api"
        }
    ],
    "paths": {
...
        "/orders": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Place an order",
                "description": "Guest checkout. Prices are computed on the server and stock is decremented inside a transaction. Throttled to 5 requests per minute.",
                "operationId": "445bab4f7dc9ea6fcd54f286c0209478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed, an item is unavailable, or there is not enough stock."
                    },
                    "429": {
                        "description": "Too many requests."
                    }
                }
            }
        },
        "/orders/track": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Track an order",
                "description": "Returns the status, status timeline, items and totals of an order when the phone and order number both match. Every mismatch returns the same generic 404. Throttled to 5 requests per minute.",
                "operationId": "04ded3f72909c0c5199696d439ad818b",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Algerian mobile number (0[5-7]XXXXXXXX) used when the order was placed.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0550123456"
                        }
                    },
                    {
                        "name": "order_number",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "BC-20261007-AB12"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The tracked order.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrackedOrder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No order matches the given phone and order number."
                    },
                    "422": {
                        "description": "The phone or order number is missing or malformed."
                    },
                    "429": {
                        "description": "Too many requests."
                    }
                }
            }
        },
        "/admin/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "List orders (admin)",
                "description": "Paginated list of the newest orders, optionally filtered by status. Requires the X-Admin-Token header.",
                "operationId": "475e4a227d3540fb0c62df6f58c47e37",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Only return orders in this status.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "shipped",
                                "delivered",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of orders.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The X-Admin-Token header is missing or does not match the configured admin token."
                    },
                    "422": {
                        "description": "The status filter is not a valid order status."
                    }
                },
                "security": [
                    {
                        "adminToken": []
                    }
                ]
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get an order (admin)",
                "description": "Returns a single order with its items and full status history. Requires the X-Admin-Token header.",
                "operationId": "0adf4cf59e05d2fd59b3e93c45235220",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The order.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The X-Admin-Token header is missing or does not match the configured admin token."
                    },
                    "404": {
                        "description": "Order not found."
                    }
                },
                "security": [
                    {
                        "adminToken": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/status": {
            "patch": {
                "tags": [
                    "orders"
                ],
                "summary": "Update an order status (admin)",
                "description": "Advances the order one step (pending → confirmed → shipped → delivered) or cancels it while it is still pending or confirmed. Cancelling restores the stock of every item. Requires the X-Admin-Token header.",
                "operationId": "65a1ea8329f5c98361011b7db6188516",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed",
                                        "enum": [
                                            "confirmed",
                                            "shipped",
                                            "delivered",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated order.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The X-Admin-Token header is missing or does not match the configured admin token."
                    },
                    "404": {
                        "description": "Order not found."
                    },
                    "422": {
                        "description": "The status is invalid or the transition is not allowed from the current status."
                    }
                },
                "security": [
                    {
                        "adminToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderRequest": {
                "required": [
                    "customer_name",
                    "phone",
                    "wilaya_id",
                    "items"
                ],
                "properties": {
                    "customer_name": {
                        "type": "string",
                        "example": "Karim Benali"
                    },
                    "phone": {
                        "description": "Algerian mobile number (0[5-7]XXXXXXXX).",
                        "type": "string",
                        "example": "0550123456"
                    },
                    "wilaya_id": {
                        "description": "ID of an active wilaya.",
                        "type": "integer",
                        "example": 16
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemRequest"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItemRequest": {
                "required": [
                    "product_id"
                ],
                "properties": {
                    "product_id": {
                        "description": "ID of an active product. Send one entry per unit ordered.",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "unit_price": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order_number": {
                        "type": "string",
                        "example": "BC-20261007-AB12"
                    },
                    "customer_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "wilaya": {
                        "type": "string"
                    },
                    "subtotal": {
                        "type": "integer"
                    },
                    "shipping_fee": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "timeline": {
                        "description": "Status history. Present only when the order history is loaded, e.g. on the admin order detail.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderStatusChange"
                        }
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrderCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "OrderStatusChange": {
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "changed_at": {
                        "type": "string",
                        "example": "2026-10-07T14:03:00.000000Z"
                    }
                },
                "type": "object"
            },
            "TrackedOrder": {
                "properties": {
                    "order_number": {
                        "type": "string",
                        "example": "BC-20261007-AB12"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "timeline": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderStatusChange"
                        }
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "subtotal": {
                        "type": "integer"
                    },
                    "shipping_fee": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "adminToken": {
                "type": "apiKey",
                "description": "Shared admin secret configured in SHOP_ADMIN_TOKEN.",
                "name": "X-Admin-Token",
                "in": "header"
            }
        }
    },
...
}